refactor: Loding 상태 처리하는 컴포넌트들 loader 폴더로 위치 변경
